+++ raphtory-graphql/src/lib.rs
@@ -1,6 +1,6 @@
 pub use crate::server::GraphServer;
 pub mod azure_auth;
-mod data;
+pub mod data;
 pub mod model;
 mod observability;
 mod routes;
